Refactor skyfield_searches/configurations.py into a package
- Decomposed `apts/skyfield_searches/configurations.py` into a modular package structure.
- Extracted `find_groupings` to `groupings.py`.
- Extracted `find_linear_alignments` to `alignments.py`.
- Hoisted inner calculation functions to top-level helpers to reduce cyclomatic complexity.
- Maintained public API via `apts/skyfield_searches/configurations/__init__.py`.
- Updated `apts/skyfield_searches/__init__.py` for submodule access.
